* changed representation of anonymous axes * made AxisCollection delegate to Axis the translation of non Axis axes (int, ...)
for the representation, we go from "axisN" (where N is the position of the axis) to "{N}".
The problem was that "axisN" was not recognizable enough as an
anonymous axis, and it was thus misleading: "a[x.axis0[...]]" would
not work.

for AxisCollection, it means all array creation functions will create
wildcard axes when given a single int for an axis labels. It also means
that stuff like this becomes possible:

ndrange(("1:5", "2:7"))
ndrange(("a,b,c", "d,e"))