made the viewer startup time on large arrays a lot faster
ndecimals and use_scientific are determined using only a tiny fraction
of the array (between 100 & 200 elements).

In practice this gave good results on all the arrays I throwed
at it. It is certainly possible to craft arrays which would be detected
wrongly, but that should be extremely rare when using real data.